fix(alert): Fix feature permissions on 'Create Alert' button (#21124)
* fixed hovercard

* style(): Auto eslint fix

* update reactnode
