+++ tests/js/spec/views/eventsV2/savedQuery/index.spec.jsx
@@ -36,7 +36,7 @@ function generateWrappedComponent(
 
 describe('EventsV2 > SaveQueryButtonGroup', function () {
   // Organization + Location does not affect state in this component
-  const organization = TestStubs.Organization();
+  const organization = TestStubs.Organization({features: ['discover-query']});
   const location = {
     pathname: '/organization/eventsv2/',
     query: {},
